Create The One
I only wanted to post this so the people that likes the selection saga could read it in this format. I'm not the author, it is Kiera Cass.
So please If you like it go and buy the originals novels, so the author could go on with the saga. 
Att. Aida